requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.37.29
botocore==1.37.29
certifi==2025.1.31
distro==1.9.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.9.0
jmespath==1.0.1
Levenshtein==0.27.1
numpy==2.2.5
openai==1.75.0
opencv-python==4.11.0.86
packaging==25.0
pillow==11.1.0
pluggy==1.5.0
pydantic==2.11.3
pydantic_core==2.33.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-Levenshtein==0.27.1
RapidFuzz==3.13.0
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.3.0
